requirements.txt
requests
python-dotenv
beautifulsoup4
whisperx
selenium
yt-dlp
audio-separator[gpu]
